ASSIGNMENT_2/requirements.txt
pandas
numpy
matplotlib.pyplot 
scipy.stats 
geopandas
altair
vega_datasets
streamlit 
zipfile
requests